% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{as_hdfql_sequence}
\alias{as_hdfql_sequence}
\title{Format Sequence For HDFql}
\usage{
as_hdfql_sequence(s)
}
\arguments{
\item{s}{An integer sequence.}
}
\value{
A vector of character representations of the integer sequence.
}
\description{
Format an integer sequence for selection with HDFql.
}
\keyword{internal}
