% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{drop}
\alias{drop}
\alias{drop_dataset}
\alias{drop_group}
\alias{drop_attribute}
\title{Drop HDF groups, datasets, and attributes}
\usage{
drop(what = c("FILE", "GROUP", "DATASET", "ATTRIBUTE"), path)

drop_dataset(dataset)

drop_group(group)

drop_attribute(attribute)
}
\arguments{
\item{what}{The type of object to drop.}

\item{path}{The target location of the object.}

\item{dataset}{The dataset to drop.}

\item{group}{The group to drop.}

\item{attribute}{The attribute to drop.}
}
\description{
Generic helper for dropping HDF objects.
}
\section{Functions}{
\itemize{
\item \code{drop_dataset}: Drop HDF dataset.

\item \code{drop_group}: Drop HDF dataset.

\item \code{drop_attribute}: Drop HDF attribute.
}}

\keyword{internal}
