% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{dtype_to_rtype}
\alias{dtype_to_rtype}
\title{HDF Data Type to R Type}
\usage{
dtype_to_rtype(dtype)
}
\arguments{
\item{dtype}{The HDF data type.}
}
\value{
The equivalent R class, or \code{NULL} if not found.
}
\description{
HDF Data Type to R Type
}
\keyword{internal}
