% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{execute_with_memory}
\alias{execute_with_memory}
\title{Execute With Memory}
\usage{
execute_with_memory(
  script,
  variable = NULL,
  direction = c("INTO", "FROM"),
  suffix = NULL,
  stop.on.error = TRUE
)
}
\arguments{
\item{script}{The HDFQL operation to execute.
Do not include \code{FROM} or \code{INTO} statements.}

\item{variable}{if not \code{NULL}, the variable to register for this operation.}

\item{direction}{Either \code{"FROM"} or \code{"INTO"}. Ignored if \code{variable}
is \code{NULL}.}

\item{suffix}{Additional script specifications. This can be used for
post-processing (for SELECT operations) or for writing raw values
(for INSERT operations).}

\item{stop.on.error}{If \code{TRUE}, return an error message if script
fails. If \code{FALSE}, return the HDFql error value.}
}
\value{
The script output, or \code{NULL}.
}
\description{
Generic helper for executing HDFql operations using memory.
}
\keyword{internal}
