% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_attr_names}
\alias{get_attr_names}
\title{Get HDF Attribute Names}
\usage{
get_attr_names(path)
}
\arguments{
\item{path}{The path of the dataset or group from which to
retrieve attribute names.}
}
\value{
A vector of attribute names.
}
\description{
Get HDF Attribute Names
}
\keyword{internal}
