% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_char_data}
\alias{get_char_data}
\title{Get Character Data}
\usage{
get_char_data(path, otype)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}
}
\value{
An R array.
}
\description{
Get character data from HDF file.
}
\keyword{internal}
