% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_charset}
\alias{get_charset}
\title{Get HDF Object Charset}
\usage{
get_charset(path, otype)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}
}
\value{
The HDF object charset.
}
\description{
Get HDF Object Charset
}
\keyword{internal}
