% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_container}
\alias{get_container}
\title{HDFql memory constructor}
\usage{
get_container(script)
}
\arguments{
\item{script}{The HDFQL operation to execute. 
Do not include `INTO` statements.}
}
\value{
An empty R object of the correct size and type.
}
\description{
Construct an R object for reading HDFql query result into memory.
}
\keyword{internal}
