% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_cursor_values}
\alias{get_cursor_values}
\title{HDFql Cursor Operation}
\usage{
get_cursor_values(script)
}
\arguments{
\item{script}{The HDFQL operation to execute. 
Do not include `INTO` statements.}
}
\value{
The script output, or `NULL`.
}
\description{
Generic helper for executing HDFql cursor operations.
}
\keyword{internal}
