% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(path, otype, transpose = TRUE)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}

\item{transpose}{If \code{TRUE}, transpose the data.}
}
\value{
An R array.
}
\description{
Get data from HDF file.
}
\keyword{internal}
