% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_data_type}
\alias{get_data_type}
\title{Get HDF Object Data Type}
\usage{
get_data_type(path, otype)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}
}
\value{
The HDF object data type.
}
\description{
Get HDF Object Data Type
}
\keyword{internal}
