% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_dimension}
\alias{get_dimension}
\title{Get HDF Object Dimension}
\usage{
get_dimension(path, otype)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}
}
\value{
The HDF object dimensions.
}
\description{
Get HDF Object Dimension
}
\keyword{internal}
