% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_key}
\alias{get_key}
\title{Get List Keys or Values}
\usage{
get_key(x, l, invert = FALSE)
}
\arguments{
\item{x}{A key or value.}

\item{l}{A named list.}

\item{invert}{If \code{TRUE}, return the key associated with
the given value.}
}
\value{
A key or value.
}
\description{
Get the value of a list key, or the key of a list value.
}
\keyword{internal}
