% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_size}
\alias{get_size}
\title{Get HDF Object Size}
\usage{
get_size(path, otype)
}
\arguments{
\item{path}{The location of the object within the HDF file.}

\item{otype}{The HDF object type.}
}
\value{
The HDF object size.
}
\description{
Get HDF Object Size
}
\keyword{internal}
