% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{get_type}
\alias{get_type}
\title{Get HDF Object Type}
\usage{
get_type(path)
}
\arguments{
\item{path}{The location of the object within the HDF file.}
}
\value{
The HDF object type.
}
\description{
Get HDF Object Type
}
\keyword{internal}
