% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\docType{data}
\name{HDFql.paths}
\alias{HDFql.paths}
\title{HDFql Library Paths}
\format{An object of class \code{environment} of length 2.}
\usage{
HDFql.paths
}
\description{
Environment containing information related to the
HDFql library on the user's system.
}
\keyword{internal}
