% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{HDFql}
\alias{HDFql}
\title{HDFql DLLs}
\description{
The HDFql DLLs are part of the Hierarchical Data Format query language (HDFql).
}
\details{
Copyright (C) 2016-2017.
For more information about HDFql, please visit the website http://www.hdfql.com.
}
