% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hdfql_check_connection}
\alias{hdfql_check_connection}
\title{Check HDFql DLLs are loaded}
\usage{
hdfql_check_connection()
}
\value{
Logical `TRUE` if DLLs are found, `FALSE` otherwise.
}
\description{
Check if the HDFql DLLs are loaded.
}
