% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hdfql_is_loaded}
\alias{hdfql_is_loaded}
\title{HDFql Library Status}
\usage{
hdfql_is_loaded()
}
\value{
Logical `TRUE` if DLLs are found, `FALSE` otherwise.
}
\description{
Check if the HDFql library loaded.
}
