% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hdfql_load}
\alias{hdfql_load}
\alias{hdfql_unload}
\title{Load HDFql DLLs}
\usage{
hdfql_load(path)

hdfql_unload()
}
\arguments{
\item{path}{The path to the HDFql installation.}
}
\description{
Load the HDFql object.
}
