% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{hdfql_operation}
\alias{hdfql_operation}
\title{HDFql Operation}
\usage{
hdfql_operation(script, retrieve = FALSE)
}
\arguments{
\item{script}{The HDFQL operation to execute. 
Do not include `INTO` statements.}

\item{retrieve}{If `TRUE`, read the operation output
into R.}
}
\description{
Generic helper for executing HDFql operations.
}
