% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\docType{data}
\name{hql.paths}
\alias{hql.paths}
\title{HDFql Library Paths}
\format{
An object of class \code{environment} of length 4.
}
\usage{
hql.paths
}
\description{
Environment containing information related to the
HDFql library on the user's system.
}
\keyword{internal}
