% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hql_attach}
\alias{hql_attach}
\alias{hql_detach}
\title{Attach HDFql Wrapper}
\usage{
hql_attach()

hql_detach()
}
\description{
Attach the constants and functions provided by the
HDFql wrapper to the search path.
}
\details{
For more information, consult the HDFql reference
manual (\url{http://www.hdfql.com/#documentation}).
}
