% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.r
\name{hql_flush}
\alias{hql_flush}
\title{Flush HDF Files}
\usage{
hql_flush(global = TRUE)
}
\arguments{
\item{global}{If \code{TRUE}, a global flush is performed and
and all open HDF files are flushed. If \code{FALSE}, a local
flush is performed and only the HDF file currently in use
is flushed.}
}
\description{
Flush HDF file(s) to write buffered data to the disk.
}
