% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hql_is_loaded}
\alias{hql_is_loaded}
\title{HDFql Library Status}
\usage{
hql_is_loaded()
}
\value{
Logical \code{TRUE} if DLLs are found, \code{FALSE} otherwise.
}
\description{
Check if the HDFql library loaded.
}
