% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{hql_load}
\alias{hql_load}
\alias{hql_unload}
\title{Load HDFql DLLs}
\usage{
hql_load(path)

hql_unload()
}
\arguments{
\item{path}{The path to the HDFql installation.}
}
\description{
Load the HDFql library.
}
\section{Functions}{
\itemize{
\item \code{hql_unload}: Unload HDFql Library.
}}

