% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{hql_read}
\alias{hql_read}
\alias{hql_read_dataset}
\alias{hql_read_attribute}
\alias{hql_read_all_attributes}
\alias{hql_read_compound_dataset}
\title{Read HDF Dataset or Attribute}
\usage{
hql_read_dataset(file, path, include.attributes = TRUE,
  parallel = FALSE)

hql_read_attribute(file, path, parallel = FALSE)

hql_read_all_attributes(file, path, parallel = FALSE)

hql_read_compound_dataset(file, path, parallel = FALSE)
}
\arguments{
\item{file}{The HDF file to read from.}

\item{path}{The location of the dataset, attribute, or group within the HDF file.}

\item{include.attributes}{If \code{TRUE}, include the dataset attributes.}

\item{parallel}{Use parallel processing functionality.}
}
\value{
A matrix.

The attribute value.

A named list of attributes.

A data frame.
}
\description{
Read a dataset or attribute from an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_read_dataset}: Read a dataset from an HDF file.

\item \code{hql_read_attribute}: Read a single attribute from an HDF file.

\item \code{hql_read_all_attributes}: Read attributes from an HDF file.

\item \code{hql_read_compound_dataset}: Read compound dataset from an HDF file.
}}

