% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{hql_use_file}
\alias{hql_use_file}
\alias{hql_close_file}
\title{Use HDF File}
\usage{
hql_use_file(file)

hql_close_file(file, all = FALSE)
}
\arguments{
\item{file}{The HDF file path.}

\item{all}{If \code{TRUE}, close all open HDF files.}
}
\description{
Open (use) an HDF file.
}
