% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{hql_write}
\alias{hql_write}
\alias{hql_write_dataset}
\alias{hql_write_attribute}
\alias{hql_write_compound_dataset}
\title{Write HDF Dataset or Attribute}
\usage{
hql_write_dataset(x, file, path)

hql_write_attribute(x, file, path)

hql_write_compound_dataset(x, file, path)
}
\arguments{
\item{x}{The dataset or attribute to write.}

\item{file}{The HDF file to write to.}

\item{path}{The location within the HDF5 file to write the dataset or attribute(s).}
}
\description{
Read a dataset or attribute to an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_write_dataset}: Write a dataset to an HDF file.

\item \code{hql_write_attribute}: Write attributes to an HDF file.

\item \code{hql_write_compound_dataset}: Write a compound dataset to an HDF file.
}}

