% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{int_to_char}
\alias{int_to_char}
\title{HDF Integer Output to Character}
\usage{
int_to_char(x, trim = FALSE)
}
\arguments{
\item{x}{An integer array.}

\item{trim}{If \code{TRUE}, trim whitespace from the character data.}
}
\value{
A character array.
}
\description{
Convert integer data from an HDF file to characters.
}
\keyword{internal}
