% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{load_hdfql}
\alias{load_hdfql}
\alias{unload_hdfql}
\title{Load HDFql DLLs}
\usage{
load_hdfql(path)

unload_hdfql()
}
\arguments{
\item{path}{The path to the HDFql installation.}
}
\description{
Load the HDFql object.
}
