% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{path_from_options}
\alias{path_from_options}
\title{HDFql Default Path}
\usage{
path_from_options(startup = FALSE)
}
\arguments{
\item{startup}{If \code{TRUE}, indicates the paths are being
detected as part of package startup.}
}
\description{
Retrieve the HDFql installation directory from existing options.
This function is used to automatically connect to HDFql without
needing to specify the installation path.
}
\details{
The function first looks for the R option \code{hdfqlr.dir},
and second looks for the environment variable \code{HDFQL-DiR}.
}
\keyword{internal}
