% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query}
\alias{query}
\title{HDFql query}
\usage{
query(script, retrieve = FALSE)
}
\arguments{
\item{script}{The HDFQL query to execute. Do not include `INTO` statements.}
}
\description{
Generic helper for executing HDFql queries.
}
