% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_attributes}
\alias{read_attributes}
\title{Read Attributes}
\usage{
read_attributes(file, path)
}
\arguments{
\item{file}{The HDF5 file path.}

\item{path}{The dataset location within the HDF5 file.}
}
\value{
A named list of attributes.
}
\description{
Read attributes of a dataset within an HDF5 file.
}
