% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_dataset}
\alias{read_dataset}
\title{Read Dataset}
\usage{
read_dataset(file, path, attributes = FALSE)
}
\arguments{
\item{file}{The HDF5 file path.}

\item{path}{The dataset location within the HDF5 file.}

\item{attributes}{If `TRUE`, include the dataset attributes.}
}
\value{
A matrix.
}
\description{
Read a dataset from an HDF5 file.
}
