% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{recurse_groups}
\alias{recurse_groups}
\alias{rev_recurse_groups}
\title{Recursively Walk Group Structure}
\usage{
recurse_groups(path)

rev_recurse_groups(file, path)
}
\description{
Recursively walk through HDF groups.
}
\keyword{internal}
