% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{rtype_to_dtype}
\alias{rtype_to_dtype}
\title{R Type to HDF Data Type}
\usage{
rtype_to_dtype(rtype, stop.on.error = TRUE)
}
\arguments{
\item{rtype}{The R class.}
}
\value{
The equivalent HDF data type, or \code{NULL} if not found.
}
\description{
R Type to HDF Data Type
}
\keyword{internal}
