% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create}
\alias{create}
\title{Create HDF Object}
\usage{
create(what = c("FILE", "GROUP", "DATASET", "ATTRIBUTE"), path, type,
  size, overwrite = FALSE, parallel = FALSE)
}
\arguments{
\item{size}{The size (dimensions) of the dataset. For \code{CHAR}
datasets, the last element of \code{size} is the string length.}

\item{overwrite}{If \code{TRUE}, overwrite existing file.}

\item{parallel}{If \code{TRUE}, use parallel capabilities.}
}
\description{
Generic helper for creating HDF objects.
}
\keyword{internal}
