% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{drop}
\alias{drop}
\alias{hql_drop_dataset}
\alias{hql_drop_group}
\alias{hql_drop_attribute}
\title{Drop HDF objects.}
\usage{
drop(what = c("FILE", "GROUP", "DATASET", "ATTRIBUTE"), path)

hql_drop_dataset(file, dataset)

hql_drop_group(file, group, recursive = FALSE)

hql_drop_attribute(file, attribute)
}
\arguments{
\item{file}{The HDF file to search.}

\item{dataset}{The dataset to drop.}

\item{group}{The group to drop.}

\item{recursive}{If \code{TRUE}, drop all child groups and datasets.}

\item{attribute}{The attribute to drop.}
}
\description{
Drop HDF objects.
}
\section{Functions}{
\itemize{
\item \code{hql_drop_dataset}: Drop HDF dataset.

\item \code{hql_drop_group}: Drop HDF group.

\item \code{hql_drop_attribute}: Drop HDF attribute.
}}

\keyword{internal}
