% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{hql_create}
\alias{hql_create}
\alias{hql_create_file}
\alias{hql_create_group}
\title{Create HDF files, groups, datasets, and attributes}
\usage{
hql_create_file(file, truncate = FALSE, parallel = FALSE,
  from = NULL, to = NULL)

hql_create_group(file, group, truncate = FALSE, object.params = list(),
  attribute.params = list())
}
\arguments{
\item{file}{The target file location.}

\item{truncate}{If \code{TRUE}, overwrite existing file.}

\item{parallel}{If \code{TRUE}, use parallel capabilities.}

\item{from, }{to Library bounds.}

\item{group}{The group to create.}

\item{file}{The file to create the group in.}

\item{truncate}{If \code{TRUE}, overwrite existing groups.}

\item{parallel}{If \code{TRUE}, use parallel capabilities.}

\item{from, }{to Library bounds.}
}
\description{
Create HDF files, groups, datasets, and attributes
}
\section{Functions}{
\itemize{
\item \code{hql_create_file}: Create HDF file.

\item \code{hql_create_group}: Create HDF group.
}}

