% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{hql_create}
\alias{hql_create}
\alias{hql_create_file}
\alias{hql_create_group}
\alias{hql_create_dataset}
\alias{hql_create_attribute}
\title{Create HDF files, groups, datasets, and attributes}
\usage{
hql_create_file(file, overwrite = FALSE, parallel = FALSE)

hql_create_group(group, overwrite = FALSE)

hql_create_dataset(dataset, data.type, size = NULL, overwrite = FALSE,
  parallel = FALSE)

hql_create_attribute(attribute, data.type, size = NULL,
  overwrite = FALSE, parallel = FALSE)
}
\arguments{
\item{file}{The target file location.}

\item{overwrite}{If \code{TRUE}, overwrite existing file.}

\item{parallel}{If \code{TRUE}, use parallel capabilities.}

\item{group}{The group to create.}

\item{dataset}{The dataset to create.}

\item{data.type}{The HDF data type of the dataset.}

\item{size}{The size (dimensions) of the dataset.}

\item{attribute}{The attribute to create.}

\item{overwrite}{If \code{TRUE}, overwrite existing groups.}

\item{data.type}{The HDF data type of the attribute.}

\item{size}{The size (dimensions) of the attribute.}
}
\description{
Create HDF files, groups, datasets, and attributes
}
\section{Functions}{
\itemize{
\item \code{hql_create_file}: Create HDF file.

\item \code{hql_create_group}: Create HDF group.

\item \code{hql_create_dataset}: Create HDF dataset.

\item \code{hql_create_attribute}: Create HDF attribute.
}}

