% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{hql_drop}
\alias{hql_drop}
\title{Drop HDF groups, datasets, and attributes}
\description{
Drop a datset, attribute, or group from an HDF file.
}
\examples{
if(hql_is_loaded()){
  tf = tempfile(fileext = ".h5")
  x = rnorm(10)
  attr(x, "myattribute") = "some information"
  hql_write_dataset(x, tf, "mygroup/mydataset")

  hql_drop_attribute(tf, "mygroup/mydataset/myattribute")
  hql_drop_dataset(tf, "mygroup/mydataset")
  hql_drop_group(tf, "mygroup")
}

}
