% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{hql_list}
\alias{hql_list}
\alias{hql_list_groups}
\alias{hql_list_datasets}
\alias{hql_list_attributes}
\title{List HDF Groups, Datasets or Attributes}
\usage{
hql_list_groups(file, path, recursive = FALSE)

hql_list_datasets(file, path, recursive = TRUE)

hql_list_attributes(file, path)
}
\arguments{
\item{file}{The HDF file to search.}

\item{path}{The location of the dataset, attribute, or group
within the HDF file.}

\item{recursive}{Recursively list child groups or datasets.}
}
\value{
A vector of paths.
}
\description{
List groups, datasets or attribute in an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_list_groups}: List groups.

\item \code{hql_list_datasets}: List datasets.

\item \code{hql_list_attributes}: List Attributes
}}

