% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{hql_read}
\alias{hql_read}
\alias{hql_read_dataset}
\alias{hql_read_attribute}
\alias{hql_read_all_attributes}
\alias{hql_read_compound_dataset}
\title{Read HDF Dataset or Attribute}
\usage{
hql_read_dataset(file, path, include.attributes = TRUE,
  parallel = FALSE)

hql_read_attribute(file, path, parallel = FALSE)

hql_read_all_attributes(file, path, parallel = FALSE)

hql_read_compound_dataset(file, path, parallel = FALSE)
}
\arguments{
\item{file}{The HDF file to read from.}

\item{path}{The location of the dataset, attribute, or group within the HDF file.}

\item{include.attributes}{If \code{TRUE}, include the dataset attributes.}

\item{parallel}{Use parallel processing functionality.}
}
\value{
A matrix.

The attribute value.

A named list of attributes.

A data frame.
}
\description{
Read a dataset or attribute from an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_read_dataset}: Read a dataset from an HDF file.

\item \code{hql_read_attribute}: Read a single attribute from an HDF file.

\item \code{hql_read_all_attributes}: Read attributes from an HDF file.

\item \code{hql_read_compound_dataset}: Read compound dataset from an HDF file.
}}

\examples{
if(hql_is_loaded()){
  tf = tempfile(fileext = ".h5")
  x = matrix(rnorm(100), nrow = 20)
  hql_write_dataset(x, tf, "dataset0")
  hql_write_attribute("normal", tf, "dataset0/dist")
 y = month.name
 attr(y, "abbreviation") = month.abb
 attr(y, "number") = 1:12
 hql_write_dataset(y, tf, "group1/dataset1")

hql_read_dataset(tf, "dataset0")
hql_read_dataset(tf, "group1/dataset1")
hql_read_attribute(tf, "group1/dataset1/abbreviation")
hql_read_all_attributes(tf, "group1/dataset1")
}

}
