% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{hql_read}
\alias{hql_read}
\alias{hql_read_dataset}
\alias{hql_read_attribute}
\alias{hql_read_all_attributes}
\title{Read HDF Dataset or Attribute}
\usage{
hql_read_dataset(file, path, attributes = FALSE)

hql_read_attribute(file, path)

hql_read_all_attributes(file, path)
}
\arguments{
\item{file}{The HDF5 file path.}

\item{path}{The location of the dataset, attribute, or group within the HDF5 file.}

\item{attributes}{If \code{TRUE}, include the dataset attributes.}
}
\value{
A matrix.

The attribute value.

A named list of attributes.
}
\description{
Read a dataset or attribute from an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_read_dataset}: Read a dataset from an HDF5 file.

\item \code{hql_read_attribute}: Read a single attribute from an HDF5 file.

\item \code{hql_read_all_attributes}: Read attributes from an HDF5 file.
}}

