% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{hql_write}
\alias{hql_write}
\alias{hql_write_dataset}
\alias{hql_write_attribute}
\alias{hql_write_all_attributes}
\alias{hql_write_compound_dataset}
\title{Write HDF Dataset or Attribute}
\usage{
hql_write_dataset(dataset, file, path, include.attributes = TRUE,
  overwrite = FALSE, parallel = FALSE)

hql_write_attribute(attribute, file, path, overwrite = FALSE,
  parallel = FALSE)

hql_write_all_attributes(attributes, file, overwrite = FALSE,
  parallel = FALSE)

hql_write_compound_dataset(dataset, file, path, overwrite = FALSE,
  parallel = FALSE)
}
\arguments{
\item{dataset}{The dataset to write.}

\item{file}{The HDF file to write to.}

\item{path}{The location within the HDF5 file to write the dataset or attribute(s).}

\item{include.attributes}{If \code{TRUE}, write the dataset attributes.}

\item{overwrite}{If \code{TRUE}, overwrite existing dataset or attribute.}

\item{parallel}{Use parallel processing functionality.}

\item{attribute}{The attribute to write.}

\item{attributes}{A list of attributes to write.}
}
\description{
Read a dataset or attribute to an HDF file.
}
\section{Functions}{
\itemize{
\item \code{hql_write_dataset}: Write a dataset to an HDF file.

\item \code{hql_write_attribute}: Write an attribute to an HDF file.

\item \code{hql_write_all_attributes}: Write multiple attributes to an HDF file.

\item \code{hql_write_compound_dataset}: Write a compound dataset to an HDF file.
}}

